Move classes and rename NamingResolvable classes to include more than SQL destinations
