+++ airbyte-integrations/bases/base-java/src/main/java/io/airbyte/integrations/destination/BufferedStreamConsumer.java
@@ -22,12 +22,13 @@
  * SOFTWARE.
  */
 
-package io.airbyte.integrations.base;
+package io.airbyte.integrations.destination;
 
 import com.google.common.base.Charsets;
 import io.airbyte.commons.concurrency.GracefulShutdownHandler;
 import io.airbyte.commons.json.Jsons;
 import io.airbyte.commons.lang.CloseableQueue;
+import io.airbyte.integrations.base.FailureTrackingConsumer;
 import io.airbyte.protocol.models.AirbyteMessage;
 import io.airbyte.protocol.models.ConfiguredAirbyteCatalog;
 import io.airbyte.queue.BigQueue;
