+++ airbyte-integrations/bases/base-java/src/main/java/io/airbyte/integrations/destination/SqlDestinationOperations.java
@@ -22,6 +22,6 @@
  * SOFTWARE.
  */
 
-package io.airbyte.integrations.base;
+package io.airbyte.integrations.destination;
 
 public interface SqlDestinationOperations extends BufferedWriteOperations, InsertTableOperations, TableCreationOperations {}
