+++ airbyte-integrations/bases/base-java/src/main/java/io/airbyte/integrations/destination/TruncateInsertIntoConsumer.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package io.airbyte.integrations.base;
+package io.airbyte.integrations.destination;
 
 import java.util.HashMap;
 import java.util.Map;
