+++ airbyte-integrations/connectors/destination-csv/src/test-integration/java/io/airbyte/integrations/destination/csv/CsvDestinationIntegrationTest.java
@@ -29,7 +29,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.collect.ImmutableMap;
 import io.airbyte.commons.json.Jsons;
-import io.airbyte.integrations.base.StandardSQLNaming;
+import io.airbyte.integrations.destination.StandardNaming;
 import io.airbyte.integrations.standardtest.destination.TestDestination;
 import java.io.FileReader;
 import java.nio.file.Files;
@@ -77,7 +77,7 @@ public class CsvDestinationIntegrationTest extends TestDestination {
   protected List<JsonNode> retrieveRecords(TestDestinationEnv testEnv, String streamName) throws Exception {
     final List<Path> allOutputs = Files.list(testEnv.getLocalRoot().resolve(RELATIVE_PATH)).collect(Collectors.toList());
     final Optional<Path> streamOutput =
-        allOutputs.stream().filter(path -> path.getFileName().toString().contains(new StandardSQLNaming().getRawTableName(streamName))).findFirst();
+        allOutputs.stream().filter(path -> path.getFileName().toString().contains(new StandardNaming().getRawTableName(streamName))).findFirst();
 
     assertTrue(streamOutput.isPresent(), "could not find output file for stream: " + streamName);
